+++ features/cerberusweb.kb/api/profiles/kb_category.php
@@ -28,7 +28,19 @@ class PageSection_ProfilesKbCategory extends Extension_PageSection {
 		Page_Profiles::renderProfile($context, $context_id, $stack);
 	}
 	
-	function savePeekJsonAction() {
+	function handleActionForPage(string $action, string $scope=null) {
+		if('profileAction' == $scope) {
+			switch ($action) {
+				case 'savePeekJson':
+					return $this->_profileAction_savePeekJson();
+				case 'viewExplore':
+					return $this->_profileAction_viewExplore();
+			}
+		}
+		return false;
+	}
+	
+	private function _profileAction_savePeekJson() {
 		@$view_id = DevblocksPlatform::importGPC($_POST['view_id'], 'string', '');
 		
 		@$id = DevblocksPlatform::importGPC($_POST['id'], 'integer', 0);
@@ -46,6 +58,12 @@ class PageSection_ProfilesKbCategory extends Extension_PageSection {
 				if(!$active_worker->hasPriv(sprintf("contexts.%s.delete", CerberusContexts::CONTEXT_KB_CATEGORY)))
 					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.delete'));
 				
+				if(false == ($model = DAO_KbCategory::get($id)))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.record.not_found'));
+				
+				if(!Context_KbCategory::isDeletableByActor($model, $active_worker))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.delete'));
+				
 				DAO_KbCategory::delete($id);
 				
 				echo json_encode(array(
@@ -58,7 +76,6 @@ class PageSection_ProfilesKbCategory extends Extension_PageSection {
 			} else {
 				@$name = DevblocksPlatform::importGPC($_POST['name'], 'string', '');
 				@$parent_id = DevblocksPlatform::importGPC($_POST['parent_id'], 'integer', 0);
-				
 				
 				if($parent_id) {
 					if(false == ($parent = DAO_KbCategory::get($parent_id)))
@@ -143,11 +160,14 @@ class PageSection_ProfilesKbCategory extends Extension_PageSection {
 	
 	}
 	
-	function viewExploreAction() {
+	private function _profileAction_viewExplore() {
 		@$view_id = DevblocksPlatform::importGPC($_POST['view_id'],'string');
 		
 		$active_worker = CerberusApplication::getActiveWorker();
 		$url_writer = DevblocksPlatform::services()->url();
+		
+		if('POST' != DevblocksPlatform::getHttpMethod())
+			DevblocksPlatform::dieWithHttpError(null, 405);
 		
 		// Generate hash
 		$hash = md5($view_id.$active_worker->id.time());
