+++ plugins/wgm.jira/api/profile/jira_issue.php
@@ -28,7 +28,19 @@ class PageSection_ProfilesJiraIssue extends Extension_PageSection {
 		Page_Profiles::renderProfile($context, $context_id, $stack);
 	}
 	
-	function savePeekAction() {
+	function handleActionForPage(string $action, string $scope=null) {
+		if('profileAction' == $scope) {
+			switch ($action) {
+				case 'savePeekJson':
+					return $this->_profileAction_savePeekJson();
+				case 'viewExplore':
+					return $this->_profileAction_viewExplore();
+			}
+		}
+		return false;
+	}
+	
+	private function _profileAction_savePeekJson() {
 		@$view_id = DevblocksPlatform::importGPC($_POST['view_id'], 'string', '');
 		
 		@$id = DevblocksPlatform::importGPC($_POST['id'], 'integer', 0);
@@ -37,23 +49,39 @@ class PageSection_ProfilesJiraIssue extends Extension_PageSection {
 		
 		$active_worker = CerberusApplication::getActiveWorker();
 		
-		if(!empty($id) && !empty($do_delete)) { // Delete
-			// [TODO] Privs
-			DAO_JiraIssue::delete($id);
+		try {
+			if (!empty($id) && !empty($do_delete)) { // Delete
+				if (!$active_worker->hasPriv(sprintf("contexts.%s.delete", Context_JiraIssue::ID)))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.delete'));
+				
+				if (false == ($model = DAO_JiraIssue::get($id)))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.record.not_found'));
+				
+				if (!Context_JiraIssue::isDeletableByActor($model, $active_worker))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.delete'));
+				
+				DAO_JiraIssue::delete($id);
+				
+			} else {
+				// Custom field saves
+				@$field_ids = DevblocksPlatform::importGPC($_POST['field_ids'], 'array', []);
+				if (!DAO_CustomFieldValue::handleFormPost(Context_JiraIssue::ID, $id, $field_ids, $error))
+					throw new Exception_DevblocksAjaxValidationError($error);
+			}
 			
-		} else {
-			// Custom field saves
-			@$field_ids = DevblocksPlatform::importGPC($_POST['field_ids'], 'array', []);
-			if(!DAO_CustomFieldValue::handleFormPost(Context_JiraIssue::ID, $id, $field_ids, $error))
-				throw new Exception_DevblocksAjaxValidationError($error);
+		} catch (Exception $e) {
+		
 		}
 	}
 	
-	function viewExploreAction() {
+	private function _profileAction_viewExplore() {
 		@$view_id = DevblocksPlatform::importGPC($_POST['view_id'],'string');
 		
 		$active_worker = CerberusApplication::getActiveWorker();
 		$url_writer = DevblocksPlatform::services()->url();
+		
+		if('POST' != DevblocksPlatform::getHttpMethod())
+			DevblocksPlatform::dieWithHttpError(null, 405);
 		
 		// Generate hash
 		$hash = md5($view_id.$active_worker->id.time());
